Move CoreText to src/core
