+++ src/components/Heading/Heading.protractor.driver.ts
@@ -1,1 +1,1 @@
-export {textDriverFactory as headingDriverFactory, TextDriver as HeadingDriver} from '../StylableUIText/CoreText/Text.protractor.driver';
+export {textDriverFactory as headingDriverFactory, TextDriver as HeadingDriver} from '../core/CoreText/Text.protractor.driver';
