+++ src/components/Heading/Heading.st.css
@@ -1,10 +1,10 @@
 :import {
-  -st-from: "../StylableUIText/CoreText/TextStyle.st.css";
+  -st-from: "../core/CoreText/TextStyle.st.css";
   -st-default: TextStyle;
 }
 
 :import {
-  -st-from: "../StylableUIText/CoreText/Text.st.css";
+  -st-from: "../core/CoreText/Text.st.css";
   -st-default: Text;
 }
