+++ src/components/StylableUIText/UIText.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {oneOf, bool, string} from 'prop-types';
-import {Text as CoreText, TextProps as CoreTextProps} from './CoreText';
+import {Text as CoreText, TextProps as CoreTextProps} from '../core/CoreText';
 import style from './UIText.st.css';
 import {withStylable} from 'wix-ui-core/withStylable';
