+++ src/components/Text/Text.driver.ts
@@ -1,5 +1,5 @@
 import { ComponentFactory } from 'wix-ui-test-utils/driver-factory';
-import { textDriverFactory as coreTextDriverFactory, TextDriver as CoreTextDriver } from '../StylableUIText/CoreText/Text.driver';
+import { textDriverFactory as coreTextDriverFactory, TextDriver as CoreTextDriver } from '../core/CoreText/Text.driver';
 import { StylableDOMUtil } from 'stylable/test-utils';
 import style from './Text.st.css';
 import { Skin, Size } from './constants';
