+++ src/components/Text/Text.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import {oneOf, Requireable} from 'prop-types';
-import {Text as CoreText, TextProps as CoreTextProps} from '../StylableUIText/CoreText';
+import {Text as CoreText, TextProps as CoreTextProps} from '../core/CoreText';
 import {withStylable} from 'wix-ui-core/withStylable';
 import {Skin, Size} from './constants';
 import style from './Text.st.css';
