+++ src/components/core/CoreText/Text.protractor.driver.ts
