+++ src/components/core/CoreText/Text.spec.tsx
