+++ src/components/core/CoreText/Text.tsx
