+++ src/components/core/CoreText/index.ts
